+++ ChangeLog.md
@@ -1,3 +1,4 @@
 # 0.3.0
 - Added YAML support to Get-ScriptVariables
 - Added new cmdlet to generate passphrases
+- Added new cmdlet to write additional log details for PSU